# monster/api/user/v1/user_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.2
// - protoc             v7.34.1
// source: user/v1/user.proto

package v1

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationUserBindPhone = "/user.v1.User/BindPhone"
const OperationUserGetDepositInfo = "/user.v1.User/GetDepositInfo"
const OperationUserGetProfile = "/user.v1.User/GetProfile"
const OperationUserGetWallet = "/user.v1.User/GetWallet"
const OperationUserListTransactions = "/user.v1.User/ListTransactions"
const OperationUserLogin = "/user.v1.User/Login"
const OperationUserLoginPhone = "/user.v1.User/LoginPhone"
const OperationUserLogout = "/user.v1.User/Logout"
const OperationUserPayDeposit = "/user.v1.User/PayDeposit"
const OperationUserRechargeDeposit = "/user.v1.User/RechargeDeposit"
const OperationUserRefundDeposit = "/user.v1.User/RefundDeposit"
const OperationUserSendSmsCode = "/user.v1.User/SendSmsCode"
const OperationUserUpdateProfile = "/user.v1.User/UpdateProfile"

type UserHTTPServer interface {
	// BindPhone 绑定手机号
	BindPhone(context.Context, *BindPhoneRequest) (*BindPhoneResponse, error)
	// GetDepositInfo 查询押金规则和缴纳状态
	GetDepositInfo(context.Context, *GetDepositInfoRequest) (*GetDepositInfoResponse, error)
	// GetProfile 获取个人信息
	GetProfile(context.Context, *GetProfileRequest) (*GetProfileResponse, error)
	// GetWallet 查询钱包资产（余额、押金、累计消费）
	GetWallet(context.Context, *GetWalletRequest) (*GetWalletResponse, error)
	// ListTransactions 查询资金流水账单
	ListTransactions(context.Context, *ListTransactionsRequest) (*ListTransactionsResponse, error)
	// Login 微信小程序登录
	Login(context.Context, *LoginRequest) (*LoginResponse, error)
	// LoginPhone 手机号验证码登录
	LoginPhone(context.Context, *LoginPhoneRequest) (*LoginResponse, error)
	// Logout 退出登录
	Logout(context.Context, *LogoutRequest) (*LogoutResponse, error)
	// PayDeposit 缴纳押金
	PayDeposit(context.Context, *PayDepositRequest) (*PayDepositResponse, error)
	// RechargeDeposit 押金快速充值
	RechargeDeposit(context.Context, *RechargeDepositRequest) (*RechargeDepositResponse, error)
	// RefundDeposit 退还押金
	RefundDeposit(context.Context, *RefundDepositRequest) (*RefundDepositResponse, error)
	// SendSmsCode 发送短信验证码
	SendSmsCode(context.Context, *SendSmsCodeRequest) (*SendSmsCodeResponse, error)
	// UpdateProfile 更新个人信息
	UpdateProfile(context.Context, *UpdateProfileRequest) (*UpdateProfileResponse, error)
}

func RegisterUserHTTPServer(s *http.Server, srv UserHTTPServer) {
	r := s.Route("/")
	r.POST("/api/v1/user/login", _User_Login0_HTTP_Handler(srv))
	r.POST("/api/v1/user/login-phone", _User_LoginPhone0_HTTP_Handler(srv))
	r.POST("/api/v1/user/send-sms-code", _User_SendSmsCode0_HTTP_Handler(srv))
	r.POST("/api/v1/user/bind-phone", _User_BindPhone0_HTTP_Handler(srv))
	r.GET("/api/v1/user/profile", _User_GetProfile0_HTTP_Handler(srv))
	r.PUT("/api/v1/user/profile", _User_UpdateProfile0_HTTP_Handler(srv))
	r.POST("/api/v1/user/logout", _User_Logout0_HTTP_Handler(srv))
	r.GET("/api/v1/wallet", _User_GetWallet0_HTTP_Handler(srv))
	r.GET("/api/v1/wallet/transactions", _User_ListTransactions0_HTTP_Handler(srv))
	r.POST("/api/v1/wallet/deposit/pay", _User_PayDeposit0_HTTP_Handler(srv))
	r.POST("/api/v1/wallet/deposit/recharge", _User_RechargeDeposit0_HTTP_Handler(srv))
	r.POST("/api/v1/wallet/deposit/refund", _User_RefundDeposit0_HTTP_Handler(srv))
	r.GET("/api/v1/wallet/deposit-info", _User_GetDepositInfo0_HTTP_Handler(srv))
}

func _User_Login0_HTTP_Handler(srv UserHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in LoginRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationUserLogin)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Login(ctx, req.(*LoginRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*LoginResponse)
		return ctx.Result(200, reply)
	}
}

func _User_LoginPhone0_HTTP_Handler(srv UserHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in LoginPhoneRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationUserLoginPhone)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.LoginPhone(ctx, req.(*LoginPhoneRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*LoginResponse)
		return ctx.Result(200, reply)
	}
}

func _User_SendSmsCode0_HTTP_Handler(srv UserHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in SendSmsCodeRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationUserSendSmsCode)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.SendSmsCode(ctx, req.(*SendSmsCodeRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*SendSmsCodeResponse)
		return ctx.Result(200, reply)
	}
}

func _User_BindPhone0_HTTP_Handler(srv UserHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in BindPhoneRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationUserBindPhone)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.BindPhone(ctx, req.(*BindPhoneRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*BindPhoneResponse)
		return ctx.Result(200, reply)
	}
}

func _User_GetProfile0_HTTP_Handler(srv UserHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetProfileRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationUserGetProfile)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetProfile(ctx, req.(*GetProfileRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetProfileResponse)
		return ctx.Result(200, reply)
	}
}

func _User_UpdateProfile0_HTTP_Handler(srv UserHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in UpdateProfileRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationUserUpdateProfile)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.UpdateProfile(ctx, req.(*UpdateProfileRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*UpdateProfileResponse)
		return ctx.Result(200, reply)
	}
}

func _User_Logout0_HTTP_Handler(srv UserHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in LogoutRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationUserLogout)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Logout(ctx, req.(*LogoutRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*LogoutResponse)
		return ctx.Result(200, reply)
	}
}

func _User_GetWallet0_HTTP_Handler(srv UserHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetWalletRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationUserGetWallet)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetWallet(ctx, req.(*GetWalletRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetWalletResponse)
		return ctx.Result(200, reply)
	}
}

func _User_ListTransactions0_HTTP_Handler(srv UserHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ListTransactionsRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationUserListTransactions)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ListTransactions(ctx, req.(*ListTransactionsRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ListTransactionsResponse)
		return ctx.Result(200, reply)
	}
}

func _User_PayDeposit0_HTTP_Handler(srv UserHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in PayDepositRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationUserPayDeposit)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.PayDeposit(ctx, req.(*PayDepositRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*PayDepositResponse)
		return ctx.Result(200, reply)
	}
}

func _User_RechargeDeposit0_HTTP_Handler(srv UserHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in RechargeDepositRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationUserRechargeDeposit)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.RechargeDeposit(ctx, req.(*RechargeDepositRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*RechargeDepositResponse)
		return ctx.Result(200, reply)
	}
}

func _User_RefundDeposit0_HTTP_Handler(srv UserHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in RefundDepositRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationUserRefundDeposit)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.RefundDeposit(ctx, req.(*RefundDepositRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*RefundDepositResponse)
		return ctx.Result(200, reply)
	}
}

func _User_GetDepositInfo0_HTTP_Handler(srv UserHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetDepositInfoRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationUserGetDepositInfo)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetDepositInfo(ctx, req.(*GetDepositInfoRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetDepositInfoResponse)
		return ctx.Result(200, reply)
	}
}

type UserHTTPClient interface {
	// BindPhone 绑定手机号
	BindPhone(ctx context.Context, req *BindPhoneRequest, opts ...http.CallOption) (rsp *BindPhoneResponse, err error)
	// GetDepositInfo 查询押金规则和缴纳状态
	GetDepositInfo(ctx context.Context, req *GetDepositInfoRequest, opts ...http.CallOption) (rsp *GetDepositInfoResponse, err error)
	// GetProfile 获取个人信息
	GetProfile(ctx context.Context, req *GetProfileRequest, opts ...http.CallOption) (rsp *GetProfileResponse, err error)
	// GetWallet 查询钱包资产（余额、押金、累计消费）
	GetWallet(ctx context.Context, req *GetWalletRequest, opts ...http.CallOption) (rsp *GetWalletResponse, err error)
	// ListTransactions 查询资金流水账单
	ListTransactions(ctx context.Context, req *ListTransactionsRequest, opts ...http.CallOption) (rsp *ListTransactionsResponse, err error)
	// Login 微信小程序登录
	Login(ctx context.Context, req *LoginRequest, opts ...http.CallOption) (rsp *LoginResponse, err error)
	// LoginPhone 手机号验证码登录
	LoginPhone(ctx context.Context, req *LoginPhoneRequest, opts ...http.CallOption) (rsp *LoginResponse, err error)
	// Logout 退出登录
	Logout(ctx context.Context, req *LogoutRequest, opts ...http.CallOption) (rsp *LogoutResponse, err error)
	// PayDeposit 缴纳押金
	PayDeposit(ctx context.Context, req *PayDepositRequest, opts ...http.CallOption) (rsp *PayDepositResponse, err error)
	// RechargeDeposit 押金快速充值
	RechargeDeposit(ctx context.Context, req *RechargeDepositRequest, opts ...http.CallOption) (rsp *RechargeDepositResponse, err error)
	// RefundDeposit 退还押金
	RefundDeposit(ctx context.Context, req *RefundDepositRequest, opts ...http.CallOption) (rsp *RefundDepositResponse, err error)
	// SendSmsCode 发送短信验证码
	SendSmsCode(ctx context.Context, req *SendSmsCodeRequest, opts ...http.CallOption) (rsp *SendSmsCodeResponse, err error)
	// UpdateProfile 更新个人信息
	UpdateProfile(ctx context.Context, req *UpdateProfileRequest, opts ...http.CallOption) (rsp *UpdateProfileResponse, err error)
}

type UserHTTPClientImpl struct {
	cc *http.Client
}

func NewUserHTTPClient(client *http.Client) UserHTTPClient {
	return &UserHTTPClientImpl{client}
}

// BindPhone 绑定手机号
func (c *UserHTTPClientImpl) BindPhone(ctx context.Context, in *BindPhoneRequest, opts ...http.CallOption) (*BindPhoneResponse, error) {
	var out BindPhoneResponse
	pattern := "/api/v1/user/bind-phone"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationUserBindPhone))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// GetDepositInfo 查询押金规则和缴纳状态
func (c *UserHTTPClientImpl) GetDepositInfo(ctx context.Context, in *GetDepositInfoRequest, opts ...http.CallOption) (*GetDepositInfoResponse, error) {
	var out GetDepositInfoResponse
	pattern := "/api/v1/wallet/deposit-info"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationUserGetDepositInfo))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// GetProfile 获取个人信息
func (c *UserHTTPClientImpl) GetProfile(ctx context.Context, in *GetProfileRequest, opts ...http.CallOption) (*GetProfileResponse, error) {
	var out GetProfileResponse
	pattern := "/api/v1/user/profile"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationUserGetProfile))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// GetWallet 查询钱包资产（余额、押金、累计消费）
func (c *UserHTTPClientImpl) GetWallet(ctx context.Context, in *GetWalletRequest, opts ...http.CallOption) (*GetWalletResponse, error) {
	var out GetWalletResponse
	pattern := "/api/v1/wallet"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationUserGetWallet))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// ListTransactions 查询资金流水账单
func (c *UserHTTPClientImpl) ListTransactions(ctx context.Context, in *ListTransactionsRequest, opts ...http.CallOption) (*ListTransactionsResponse, error) {
	var out ListTransactionsResponse
	pattern := "/api/v1/wallet/transactions"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationUserListTransactions))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// Login 微信小程序登录
func (c *UserHTTPClientImpl) Login(ctx context.Context, in *LoginRequest, opts ...http.CallOption) (*LoginResponse, error) {
	var out LoginResponse
	pattern := "/api/v1/user/login"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationUserLogin))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// LoginPhone 手机号验证码登录
func (c *UserHTTPClientImpl) LoginPhone(ctx context.Context, in *LoginPhoneRequest, opts ...http.CallOption) (*LoginResponse, error) {
	var out LoginResponse
	pattern := "/api/v1/user/login-phone"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationUserLoginPhone))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// Logout 退出登录
func (c *UserHTTPClientImpl) Logout(ctx context.Context, in *LogoutRequest, opts ...http.CallOption) (*LogoutResponse, error) {
	var out LogoutResponse
	pattern := "/api/v1/user/logout"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationUserLogout))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// PayDeposit 缴纳押金
func (c *UserHTTPClientImpl) PayDeposit(ctx context.Context, in *PayDepositRequest, opts ...http.CallOption) (*PayDepositResponse, error) {
	var out PayDepositResponse
	pattern := "/api/v1/wallet/deposit/pay"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationUserPayDeposit))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// RechargeDeposit 押金快速充值
func (c *UserHTTPClientImpl) RechargeDeposit(ctx context.Context, in *RechargeDepositRequest, opts ...http.CallOption) (*RechargeDepositResponse, error) {
	var out RechargeDepositResponse
	pattern := "/api/v1/wallet/deposit/recharge"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationUserRechargeDeposit))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// RefundDeposit 退还押金
func (c *UserHTTPClientImpl) RefundDeposit(ctx context.Context, in *RefundDepositRequest, opts ...http.CallOption) (*RefundDepositResponse, error) {
	var out RefundDepositResponse
	pattern := "/api/v1/wallet/deposit/refund"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationUserRefundDeposit))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// SendSmsCode 发送短信验证码
func (c *UserHTTPClientImpl) SendSmsCode(ctx context.Context, in *SendSmsCodeRequest, opts ...http.CallOption) (*SendSmsCodeResponse, error) {
	var out SendSmsCodeResponse
	pattern := "/api/v1/user/send-sms-code"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationUserSendSmsCode))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// UpdateProfile 更新个人信息
func (c *UserHTTPClientImpl) UpdateProfile(ctx context.Context, in *UpdateProfileRequest, opts ...http.CallOption) (*UpdateProfileResponse, error) {
	var out UpdateProfileResponse
	pattern := "/api/v1/user/profile"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationUserUpdateProfile))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "PUT", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}