# monster/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /api/v1/user/bind-phone:
        post:
            tags:
                - User
            description: 绑定手机号
            operationId: User_BindPhone
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/user.v1.BindPhoneRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/user.v1.BindPhoneResponse'
    /api/v1/user/login:
        post:
            tags:
                - User
            description: 微信小程序登录
            operationId: User_Login
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/user.v1.LoginRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/user.v1.LoginResponse'
    /api/v1/user/login-phone:
        post:
            tags:
                - User
            description: 手机号验证码登录
            operationId: User_LoginPhone
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/user.v1.LoginPhoneRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/user.v1.LoginResponse'
    /api/v1/user/logout:
        post:
            tags:
                - User
            description: 退出登录
            operationId: User_Logout
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/user.v1.LogoutRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/user.v1.LogoutResponse'
    /api/v1/user/profile:
        get:
            tags:
                - User
            description: 获取个人信息
            operationId: User_GetProfile
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/user.v1.GetProfileResponse'
        put:
            tags:
                - User
            description: 更新个人信息
            operationId: User_UpdateProfile
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/user.v1.UpdateProfileRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/user.v1.UpdateProfileResponse'
    /api/v1/user/send-sms-code:
        post:
            tags:
                - User
            description: 发送短信验证码
            operationId: User_SendSmsCode
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/user.v1.SendSmsCodeRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/user.v1.SendSmsCodeResponse'
...
    /api/v1/wallet/deposit/pay:
        post:
            tags:
                - User
            description: 缴纳押金
            operationId: User_PayDeposit
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/user.v1.PayDepositRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/user.v1.PayDepositResponse'
    /api/v1/wallet/deposit/recharge:
        post:
            tags:
                - User
            description: 押金快速充值
            operationId: User_RechargeDeposit
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/user.v1.RechargeDepositRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/user.v1.RechargeDepositResponse'
    /api/v1/wallet/deposit/refund:
        post:
            tags:
                - User
            description: 退还押金
            operationId: User_RefundDeposit
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/user.v1.RefundDepositRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/user.v1.RefundDepositResponse'
    /api/v1/wallet/transactions:
        get:
            tags:
                - User
            description: 查询资金流水账单
            operationId: User_ListTransactions
            parameters:
                - name: page
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: pageSize
                  in: query
                  schema:
                    type: integer
                    format: int32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/user.v1.ListTransactionsResponse'
components:
    schemas:
...
        user.v1.BindPhoneRequest:
            type: object
            properties:
                phone:
                    type: string
                code:
                    type: string
        user.v1.BindPhoneResponse:
            type: object
            properties:
                message:
                    type: string
...
        user.v1.GetProfileResponse:
            type: object
            properties:
                user:
                    $ref: '#/components/schemas/user.v1.UserInfo'
...
        user.v1.ListTransactionsResponse:
            type: object
            properties:
                list:
                    type: array
                    items:
                        $ref: '#/components/schemas/user.v1.TransactionInfo'
                total:
                    type: integer
                    format: int32
                page:
                    type: integer
                    format: int32
                pageSize:
                    type: integer
                    format: int32
        user.v1.LoginPhoneRequest:
            type: object
            properties:
                phone:
                    type: string
                code:
                    type: string
                nickname:
                    type: string
                avatar:
                    type: string
        user.v1.LoginRequest:
            type: object
            properties:
                code:
                    type: string
                nickname:
                    type: string
                avatar:
                    type: string
        user.v1.LoginResponse:
            type: object
            properties:
                token:
                    type: string
                user:
                    $ref: '#/components/schemas/user.v1.UserInfo'
        user.v1.LogoutRequest:
            type: object
            properties: {}
        user.v1.LogoutResponse:
            type: object
            properties:
                message:
                    type: string
        user.v1.PayDepositRequest:
            type: object
            properties:
                amount:
                    type: string
        user.v1.PayDepositResponse:
            type: object
            properties:
                message:
                    type: string
        user.v1.RechargeDepositRequest:
            type: object
            properties:
                amount:
                    type: string
        user.v1.RechargeDepositResponse:
            type: object
            properties:
                message:
                    type: string
        user.v1.RefundDepositRequest:
            type: object
            properties: {}
        user.v1.RefundDepositResponse:
            type: object
            properties:
                message:
                    type: string
        user.v1.SendSmsCodeRequest:
            type: object
            properties:
                phone:
                    type: string
        user.v1.SendSmsCodeResponse:
            type: object
            properties:
                message:
                    type: string
        user.v1.TransactionInfo:
            type: object
            properties:
                id:
                    type: string
                userId:
                    type: string
                walletId:
                    type: string
                orderId:
                    type: string
                tradeType:
                    type: string
                amount:
                    type: string
                balanceBefore:
                    type: string
                balanceAfter:
                    type: string
                remark:
                    type: string
                createdAt:
                    type: string
        user.v1.UpdateProfileRequest:
            type: object
            properties:
                nickname:
                    type: string
                avatar:
                    type: string
                gender:
                    type: integer
                    format: int32
        user.v1.UpdateProfileResponse:
            type: object
            properties:
                user:
                    $ref: '#/components/schemas/user.v1.UserInfo'
        user.v1.UserInfo:
            type: object
            properties:
                id:
                    type: string
                openId:
                    type: string
                unionId:
                    type: string
                nickname:
                    type: string
                avatar:
                    type: string
                phone:
                    type: string
                gender:
                    type: integer
                    format: int32
                status:
                    type: integer
                    format: int32
                createdAt:
                    type: string
                updatedAt:
                    type: string
...